+++ dracoon/nodes/__init__.py
@@ -27,11 +27,12 @@ import urllib.parse
 import httpx
 from pydantic import validate_arguments
 from tqdm import tqdm
+from tenacity import retry
 
 from dracoon.crypto import FileEncryptionCipher, encrypt_bytes, encrypt_file_key, create_file_key, get_file_key_version
 from dracoon.crypto.models import FileKey, PlainUserKeyPairContainer, UserKeyPairContainer
 from dracoon.groups.models import Expiration
-from dracoon.client import DRACOONClient, OAuth2ConnectionType
+from dracoon.client import DRACOONClient, OAuth2ConnectionType, RETRY_CONFIG
 from dracoon.errors import (InvalidClientError, ClientDisconnectedError, InvalidFileError, InvalidArgumentError)
 from dracoon.uploads.models import UploadChannelResponse
 from .models import (Callback, CompleteS3Upload, CompleteUpload, ConfigRoom, CreateFolder, CreateRoom, CreateUploadChannel, EncryptRoom, 
@@ -108,7 +109,7 @@ class DRACOONNodes:
             yield data
 
     # get download url as authenticated user to download a file
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def create_upload_channel(self, upload_channel: CreateUploadChannel, raise_on_err: bool = False) -> CreateFileUploadResponse:
         """ create an upload channel to upload (S3 direct or proxy) """
         payload = upload_channel.dict(exclude_unset=True)
@@ -152,7 +153,7 @@ class DRACOONNodes:
         return CreateUploadChannel(**upload_channel)
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def cancel_upload(self, upload_id: str, raise_on_err: bool = False) -> None:
         """ cancel an upload channel (and delete chunks) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -173,7 +174,8 @@ class DRACOONNodes:
 
         self.logger.info("Upload canceled.")
         return None
-
+    
+    @retry(**RETRY_CONFIG)
     async def get_node_from_path(self, path: str, filter: str = None, raise_on_err: bool = False) -> Node:
         """ get node id from path """
         
@@ -221,7 +223,7 @@ class DRACOONNodes:
             return None
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def complete_s3_upload(self, upload_id: str, upload: CompleteS3Upload, raise_on_err: bool = False) -> None:
         """ finalize an S3 direct upload """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -275,7 +277,7 @@ class DRACOONNodes:
         
         return GetS3Urls(**payload)
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_s3_urls(self, upload_id: str, upload: GetS3Urls, raise_on_err: bool = False) -> PresignedUrlList:
         """ get a list of S3 urls based on provided chunk count """
         """ chunk size needs to be larger than 5 MB """
@@ -301,6 +303,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved S3 presigned upload URLs.")
         return PresignedUrlList(**res.json())
     
+    @retry(**RETRY_CONFIG)
     async def upload_unencrypted(self, file_path: str, upload_channel: CreateFileUploadResponse, keep_shares: bool = False, 
                                     file_name: str = None, resolution_strategy: str = 'autorename', chunksize: int = CHUNK_SIZE, 
                                     display_progress: bool = False, raise_on_err: bool = False, 
@@ -395,7 +398,7 @@ class DRACOONNodes:
              
         return node
     
-                                  
+    @retry(**RETRY_CONFIG)                     
     async def upload_encrypted(self, file_path: str, upload_channel: CreateFileUploadResponse, plain_keypair: PlainUserKeyPairContainer, 
                                   file_name: str = None, keep_shares: bool = False, resolution_strategy: str = 'autorename', 
                                   chunksize: int = CHUNK_SIZE, display_progress: bool = False, raise_on_err: bool = False,
@@ -538,7 +541,8 @@ class DRACOONNodes:
             
         
         return CompleteUpload(**payload)
-            
+    
+    @retry(**RETRY_CONFIG)   
     async def complete_upload(self, upload_channel: UploadChannelResponse, payload: CompleteUpload, raise_on_err: bool = False) -> Node:
         
         api_url = self.dracoon.base_url + self.dracoon.api_base_url + f'/uploads/{upload_channel.token}'
@@ -556,6 +560,7 @@ class DRACOONNodes:
           
         return Node(**res.json())
     
+    @retry(**RETRY_CONFIG)
     async def upload_s3_unencrypted(self, file_path: str, upload_channel: CreateFileUploadResponse, keep_shares: bool = False,
                                     file_name: str = None,
                                     resolution_strategy: str = 'autorename', chunksize: int = CHUNK_SIZE, 
@@ -707,7 +712,8 @@ class DRACOONNodes:
             time *= 2
             
         return upload_status
-        
+    
+    @retry(**RETRY_CONFIG)
     async def upload_s3_encrypted(self, file_path: str, upload_channel: CreateFileUploadResponse, plain_keypair: PlainUserKeyPairContainer, 
                                   file_name: str = None,
                                   keep_shares: bool = False, resolution_strategy: str = 'autorename', 
@@ -887,7 +893,8 @@ class DRACOONNodes:
             time *= 2
             
         return upload_status
-            
+    
+    @retry(**RETRY_CONFIG)
     async def check_s3_upload(self, upload_id: str, raise_on_err: bool = False) -> S3FileUploadStatus:
         """ check status of S3 upload """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -908,7 +915,7 @@ class DRACOONNodes:
         return S3FileUploadStatus(**res.json())
         
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_nodes(self, room_manager: bool = False, parent_id: int = 0, offset: int = 0, filter: str = None, limit: int = None, sort: str = None, 
                         raise_on_err: bool = False) -> NodeList:
         """ list (all) visible nodes """
@@ -941,8 +948,8 @@ class DRACOONNodes:
         self.logger.info("Retrieved nodes.")
         return NodeList(**res.json())
     
-    # delete nodes for given array of node ids
-    @validate_arguments
+
+    @retry(**RETRY_CONFIG) 
     async def delete_nodes(self, node_list: List[int], raise_on_err: bool = False) -> None:
         """ delete a list of nodes (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -968,8 +975,7 @@ class DRACOONNodes:
         self.logger.info("Deleted node(s).")
         return None
 
-    # get node for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG) 
     async def get_node(self, node_id: int, raise_on_err: bool = False) -> Node:
         """ get specific node details (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -993,7 +999,7 @@ class DRACOONNodes:
         return Node(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def delete_node(self, node_id: int, raise_on_err: bool = False) -> None:
         """ delete specific node (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1017,8 +1023,7 @@ class DRACOONNodes:
         self.logger.info("Deleted node.")
         return None
 
-    # get node comments for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_node_comments(self, node_id: int, offset: int = 0, raise_on_err: bool = False) -> CommentList:
         """ get comments for specific node (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1043,8 +1048,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved node comments.")
         return CommentList(**res.json())
 
-    # get node for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def add_node_comment(self, node_id: int, comment: CommentNode, raise_on_err: bool = False) -> Comment:
         """ add a comment to a node """
         payload = comment.dict(exclude_unset=True)
@@ -1077,8 +1081,7 @@ class DRACOONNodes:
 
         return CommentNode(**comment)
 
-    # copy node for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def copy_nodes(self, target_id: int, copy_node: TransferNode, raise_on_err: bool = False) -> Node:
         """ copy node(s) to given target id """
         payload = copy_node.dict(exclude_unset=True)
@@ -1128,8 +1131,7 @@ class DRACOONNodes:
         
         return NodeItem(**node_item)
 
-    # get node comfor given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_deleted_nodes(self, parent_id: int = 0, offset: int = 0, filter: str = None, limit: int = None, sort: str = None, raise_on_err: bool = False) -> DeletedNodeSummaryList:
         """ list (all) deleted nodes """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1164,7 +1166,7 @@ class DRACOONNodes:
         return DeletedNodeSummaryList(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def empty_node_recyclebin(self, parent_id: int, raise_on_err: bool = False) -> None:
         """ delete all nodes in recycle bin of parent (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1188,8 +1190,7 @@ class DRACOONNodes:
         self.logger.info("Emptied node recycle bin.")
         return None
 
-    # get node versions in a given parent id (requires name, specification of type)
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_node_versions(self, parent_id: int, name: str, type: str, offset: int = 0, raise_on_err: bool = False) -> DeletedNodeVersionsList:
         """ get (all) versions of a node (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1216,7 +1217,7 @@ class DRACOONNodes:
         return DeletedNodeVersionsList(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def add_favorite(self, node_id: int, raise_on_err: bool = False) -> Node:
         """ add a specific node to favorites (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1239,8 +1240,7 @@ class DRACOONNodes:
         self.logger.info("Added node to favorites.")
         return Node(**res.json())
 
-    # delete node for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def delete_favorite(self, node_id: int, raise_on_err: bool = False) -> None:
         """ remove a specific node from favorites (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1263,8 +1263,7 @@ class DRACOONNodes:
         self.logger.info("Removed node from favorites.")
         return None
 
-    # copy node for given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def move_nodes(self, target_id: int, move_node: TransferNode, raise_on_err: bool = False) -> Node:
         """ move node(s) to target node (by id) """
         payload = move_node.dict(exclude_unset=True)
@@ -1289,8 +1288,7 @@ class DRACOONNodes:
         return Node(**res.json())
 
 
-    # get node ancestors (parents)
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_parents(self, node_id: int, raise_on_err: bool = False) -> NodeParentList:
         """ get node parents """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1313,9 +1311,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved node parents.")
         return NodeParentList(**res.json())
 
-    # delete deleted nodes in recycle bin for given array of node ids
-
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def empty_recyclebin(self, node_list: List[int], raise_on_err: bool = False) -> None:
         """ empty recylce bin: list of nodes (deleted nodes by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1343,7 +1339,7 @@ class DRACOONNodes:
         self.logger.info("Emptied recycle bin.")
         return None
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_deleted_node(self, node_id: int, raise_on_err: bool = False) -> DeletedNode:
         """ get details of a specific deleted node (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1368,7 +1364,7 @@ class DRACOONNodes:
         return DeletedNode(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def restore_nodes(self, restore: RestoreNode, raise_on_err: bool = False) -> None:
         """ restore a list of nodes from recycle bin """
         payload = restore.dict(exclude_unset=True)
@@ -1406,8 +1402,7 @@ class DRACOONNodes:
         
         return RestoreNode(**node_restore)
 
-    # update file meta data
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_file(self, file_id: int, file_update: UpdateFile, raise_on_err: bool = False) -> Node:
         """ update file metadata """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1433,7 +1428,7 @@ class DRACOONNodes:
         self.logger.info("Updated file.")
         return Node(**res.json())
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_files(self, files_update: UpdateFiles, raise_on_err: bool = False) -> None:
         """ update file metadata """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1489,7 +1484,7 @@ class DRACOONNodes:
         return UpdateFiles(**update_files)
     
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_download_url(self, node_id: int, raise_on_err: bool = False) -> DownloadTokenGenerateResponse:
         """ get download url for a specific node """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1510,8 +1505,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved download URL.")
         return DownloadTokenGenerateResponse(**res.json())
 
-    # get user file key if available
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_user_file_key(self, file_id: int, version: str = None, raise_on_err: bool = False) -> FileKey:
         """ get file key for given node as authenticated user """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1536,7 +1530,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved user file key.")
         return FileKey(**res.json())
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def set_file_keys(self, file_keys: SetFileKeys, raise_on_err: bool = False) -> None:
         """ set file keys for nodes """
         payload = file_keys.dict(exclude_unset=True)
@@ -1576,7 +1570,8 @@ class DRACOONNodes:
             "userId": user_id, 
             "fileKey": file_key
         })
-        
+    
+    @retry(**RETRY_CONFIG)
     async def get_file_versions(self, reference_id: int, raise_on_err: bool = False):
         """ get all file versions (including deleted nodes) for given reference id """
         
@@ -1600,7 +1595,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved node.")
         return Node(**res.json())
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def create_folder(self, folder: CreateFolder, raise_on_err: bool = False) -> Node:
         """ create a new folder """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1650,8 +1645,7 @@ class DRACOONNodes:
 
         return UpdateFolder(**folder)
 
-    # update folder mets data
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_folder(self, node_id: int, folder_update: UpdateFolder, raise_on_err: bool = False) -> Node:
         """ update a folder """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1676,8 +1670,7 @@ class DRACOONNodes:
         self.logger.info("Updated folder.")
         return Node(**res.json())
 
-    # get missing file keys
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_missing_file_keys(self, file_id: int = None, room_id: int = None, user_id: int = None, use_key: str = None, 
                                       offset: int = 0, limit: int = None, raise_on_err: bool = False) -> MissingKeysResponse:
         """ get (all) missing file keys """
@@ -1713,8 +1706,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved missing file keys.")
         return MissingKeysResponse(**res.json())
 
-    # create room
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def create_room(self, room: CreateRoom, raise_on_err: bool = False) -> Node:
         """ create a new room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1740,8 +1732,7 @@ class DRACOONNodes:
         return Node(**res.json())
     
 
-    # update room mets data
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_room(self, node_id: int, room_update: UpdateRoom, raise_on_err: bool = False) -> Node:
         """ update a room (by id) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1805,8 +1796,7 @@ class DRACOONNodes:
 
         return UpdateRoom(**room)
 
-    # configure data room
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def config_room(self, node_id: int, config_update: ConfigRoom, raise_on_err: bool = False) -> Node:
         """ configure a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1879,8 +1869,7 @@ class DRACOONNodes:
             "permissions": permission
 
         }
-    
-    @validate_arguments
+
     def make_encrypt_room(self, is_encrypted: bool, use_sytem_rescue_key: bool = None, room_rescue_key: UserKeyPairContainer = None) -> EncryptRoom:
         """ make room encryption payload """
         
@@ -1893,7 +1882,7 @@ class DRACOONNodes:
         
         return encrypt_room
     
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def encrypt_room(self, room_id: int, encrypt_room: EncryptRoom, raise_on_err: bool = False) -> Node:
         
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1918,8 +1907,7 @@ class DRACOONNodes:
         self.logger.info("Encrypted room.")
         return Node(**res.json())
 
-    # get node comfor given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_room_groups(self, room_id: int, offset: int = 0, filter: str = None, limit: str = None, sort: str = None, raise_on_err: bool = False) -> RoomGroupList:
         """ list (all) groups assigned to a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1953,7 +1941,7 @@ class DRACOONNodes:
         return RoomGroupList(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_room_groups(self, room_id: int, groups_update: UpdateRoomGroups, raise_on_err: bool = False) -> None:
         """ bulk update assigned groups of a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -1979,8 +1967,7 @@ class DRACOONNodes:
         self.logger.info("Updated room groups.")
         return None
 
-    # delete groups assigned to room with given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def delete_room_groups(self, room_id: int, group_list: List[int], raise_on_err: bool = False) -> None:
         """ bulk delete assigned groups of a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2007,9 +1994,7 @@ class DRACOONNodes:
         self.logger.info("Deleted room group(s).")
         return None
 
-    # get node comfor given node id
-
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_room_users(self, room_id: int, offset: int = 0, filter: str = None, limit: str = None, sort: str = None, raise_on_err: bool = False) -> RoomUserList:
         """ get (all) users assigned to a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2042,8 +2027,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved room users.")
         return RoomUserList(**res.json())
 
-    # add or change users assigned to room with given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_room_users(self, room_id: int, users_update: UpdateRoomUsers, raise_on_err: bool = False) -> None:
         """ bulk update assigned users in a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2072,8 +2056,7 @@ class DRACOONNodes:
         self.logger.info("Updated room user(s).")
         return None
 
-    # delete users assigned to room with given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def delete_room_users(self, room_id: int, user_list: List[int], raise_on_err: bool = False) -> None:
         """ bulk remove assigned users in a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2100,8 +2083,7 @@ class DRACOONNodes:
         self.logger.info("Deleted room user(s).")
         return None
 
-    # get webhooks assigned or assignable to room with given node id
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_room_webhooks(self, node_id: int, offset: int = 0, filter: str = None, limit: str = None, sort: str = None, raise_on_err: bool = False) -> RoomWebhookList:
         """" list (all) room webhooks """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2135,9 +2117,7 @@ class DRACOONNodes:
         self.logger.info("Retrieved room webhooks.")
         return RoomWebhookList(**res.json())
 
-    # delete users assigned to room with given node id
-
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def update_room_webhooks(self, node_id: int, hook_update: UpdateRoomHooks, raise_on_err: bool = False) -> RoomWebhookList:
         """ update room webhooks """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2163,7 +2143,7 @@ class DRACOONNodes:
         self.logger.info("Updated room webhooks.")
         return RoomWebhookList(**res.json())
     
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_room_events(self, room_id: int, offset: int = 0, filter: str = None, limit: int = None, 
                         sort: str = None, date_start: str = None, date_end: str = None, operation_id: int = None, user_id: int = None, raise_on_err = False) -> LogEventList:
         """ get pending room assignments (new group members not accepted) """
@@ -2199,7 +2179,7 @@ class DRACOONNodes:
         return LogEventList(**res.json())
 
     
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def get_pending_assignments(self, offset: int = 0, filter: str = None, limit: str = None, sort: str = None, raise_on_err: bool = False) -> PendingAssignmentList:
         """ get pending room assignments (new group members not accepted) """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2233,7 +2213,7 @@ class DRACOONNodes:
         return PendingAssignmentList(**res.json())
 
 
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def process_pending_assignments(self, pending_update: ProcessRoomPendingUsers, raise_on_err: bool = False) -> None:
         """ procces (accept or reject) new group members of a room """
         if not await self.dracoon.test_connection() and self.dracoon.connection:
@@ -2259,8 +2239,7 @@ class DRACOONNodes:
         self.logger.info("Processed pending assignments.")
         return None
 
-    # search for nodes
-    @validate_arguments
+    @retry(**RETRY_CONFIG)
     async def search_nodes(self, search: str, parent_id: int = 0, depth_level: int = 0, offset: int = 0, 
                            filter: str = None, limit: str = None, sort: str = None, raise_on_err: bool = False) -> NodeList:
         """ search for specific nodes """
